Move views for bellhop-auth-dummy into the correct crate (#5)
