+++ bellhop-auth-dummy/src/lib.rs
@@ -2,6 +2,8 @@
 
 #[macro_use]
 extern crate rocket;
+#[macro_use]
+extern crate serde_derive;
 
 mod views;
 
@@ -17,7 +19,7 @@ const LOGIN_COOKIE: &str = "user_login";
 
 #[catch(401)]
 fn unauthorized() -> Redirect {
-    Redirect::to("/login")
+    Redirect::to("/auth/dummy/login")
 }
 
 #[derive(Debug)]
@@ -25,7 +27,9 @@ pub struct Dummy;
 
 impl Auth for Dummy {
     fn prelaunch(&self, rocket: Rocket) -> Rocket {
-        rocket.register(catchers![unauthorized])
+        rocket
+            .register(catchers![unauthorized])
+            .mount("/auth/dummy/", routes![views::login_get, views::logout, views::login_post])
     }
 
     fn authenticate(&self, c: &Db, req: &Request) -> Result<Option<User>, Error> {
