+++ bellhop/src/views/mod.rs
@@ -1,6 +1,5 @@
 pub mod assets;
 pub mod endpoints;
 pub mod favicon;
-pub mod login;
 pub mod types;
 pub mod user;
