+++ bellhop/templates/base.html.hbs
@@ -39,7 +39,7 @@
             <div class="pure-u-1 pure-u-md-1-3">
                 <div class="pure-menu pure-menu-horizontal custom-menu-3-right">
                     <ul class="pure-menu-list">
-                        <li class="pure-menu-item"><a href="/logout" class="pure-menu-link custom-menu-item">Logout</a></li>
+                        <li class="pure-menu-item"><a href="/auth/dummy/logout" class="pure-menu-link custom-menu-item">Logout</a></li>
                         <li class="pure-menu-item"><a class="pure-menu-link custom-menu-item">{{user.email}}</a></li>
                     </ul>
                 </div>
