+++ bellhop/templates/login/home.html.hbs
@@ -3,7 +3,7 @@
 {{/inline}}
 
 {{#*inline "content"}}
-    <form class="pure-form" action="/login" method="post">
+    <form class="pure-form" method="post">
         Email:<br>
         <input  type="text" name="email">
         <br>
